requirements.txt
# requirements.txt
streamlit==1.48.0   
google-generativeai==0.8.5 
googlemaps==4.10.0
